Use 32-bit array sizes.

+++ vespalib/src/vespa/vespalib/datastore/array_store_type_mapper.h
@@ -18,7 +18,7 @@ namespace vespalib::datastore {
 class ArrayStoreTypeMapper
 {
 protected:
-    std::vector<size_t> _array_sizes;
+    std::vector<uint32_t> _array_sizes;
 public:
     ArrayStoreTypeMapper();
     ~ArrayStoreTypeMapper();
